simplified conf, added logging to usercontroller

+++ app/controllers/UserController.java
@@ -4,6 +4,8 @@ import forms.LoginForm;
 import forms.UserPasswordForm;
 import forms.UserSettingsForm;
 import models.User;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import play.data.Form;
@@ -18,23 +20,32 @@ import views.html.signup;
 @Controller
 public class UserController extends play.mvc.Controller {
 
+    private static final Logger logger = LoggerFactory.getLogger(UserController.class);
+
     @Autowired
     private UserService userService;
 
-    /****************************************
-     * API routes
-     */
+    /* ***************************************
+     *  API routes
+     * ***************************************/
 
     @Security.Authenticated(SecuredApi.class)
     public Result deleteUserAPI(Integer userId) {
-        try {
-            userService.deleteUser(userId);
-            flash("info", "User deleted successfully");
-            session().clear();
-            return ok("User deleted");
-        } catch (NullPointerException | NumberFormatException e) {
-            flash("error", "Unable to delete user");
-            return badRequest("Error: unable to delete user");
+
+        if(!session().get("userid").equals(userId+"")) {
+            logger.error("User "+session().get("userid")+" tried to delete user "+userId+"!");
+            flash("error", "(╯°□°)╯︵ ┻━┻ YOU CAN'T HACK ME! I HACK YOU!");
+            return badRequest("(╯°□°)╯︵ ┻━┻ YOU CAN'T HACK ME! I HACK YOU!");
+        } else {
+            try {
+                userService.deleteUser(userId);
+                flash("info", "User deleted successfully");
+                session().clear();
+                return ok("User deleted");
+            } catch (NullPointerException | NumberFormatException e) {
+                flash("error", "Unable to delete user");
+                return badRequest("Error: unable to delete user");
+            }
         }
     }
 
@@ -52,21 +63,21 @@ public class UserController extends play.mvc.Controller {
         return ok();
     }
 
-    /******************************************
-     * User content routes
-     */
+    /* *****************************************
+     *  User content routes
+     * ***************************************/
 
     public Result getLogin() {
         return ok(login.render(play.data.Form.form(LoginForm.class)));
     }
 
     public Result doLogout() {
+        logger.info("User logged out: "+session().get("userid"));
         session().clear();
         return redirect(routes.Index.index());
     }
 
     public Result doLogin() {
-
         User user = null;
         Form<LoginForm> form = Form.form(LoginForm.class).bindFromRequest();
         if (!form.hasErrors()) {
@@ -80,14 +91,17 @@ public class UserController extends play.mvc.Controller {
             }
         }
         if (form.hasErrors()) {
+            logger.info("User login rejected for "+user.getUsername()+": "+form.errorsAsJson());
             flash("error", "Login error. Get that weak shit outta here.");
             return badRequest(login.render(form));
         }
 
         if (user != null) {
+            logger.info("User logged in: "+user.getId()+" ("+user.getUsername()+")");
             session("userid", "" + user.getId());
             return redirect(routes.Index.home());
         } else {
+            logger.error("User login failed! User was null :(");
             flash("error", "OH SHIT! Something broke on our side. Try that again if you want.");
             return internalServerError(login.render(form));
         }
@@ -98,7 +112,6 @@ public class UserController extends play.mvc.Controller {
     }
 
     public Result doSignup() {
-
         Form<models.User> form = Form.form(models.User.class).bindFromRequest();
         if (form.hasErrors()) {
             flash("error", "Try that again, dipshit.");
@@ -111,6 +124,7 @@ public class UserController extends play.mvc.Controller {
             return badRequest(signup.render(form));
         }
 
+        logger.info("New user created: "+newUser.getUsername());
         userService.saveUser(newUser);
         flash("success", "Welcome aboard, " + newUser.getAlias() + ". Login to continue.");
         return redirect(routes.UserController.getLogin());
